refactor(init): rename internal/init to internal/initialize
Rename the `internal/init` package to `internal/initialize` to avoid
Go reserved keyword conflict. This eliminates the need for awkward
import aliases like `initpkg` in consuming code.

Changes:
- Rename directory: internal/init/ -> internal/initialize/
- Update package declarations from `package init` to `package initialize`
- Update all import paths in executor.go, wizard.go, and cmd/init.go
- Remove unnecessary `initpkg` alias from cmd/init.go

This is a pure refactoring with no behavioral changes. All tests pass
and the build succeeds.
