+++ internal/initialize/filesystem_test.go
@@ -1,4 +1,4 @@
-package init
+package initialize
 
 import (
 	"os"
