+++ internal/initialize/providers/registry_test.go
