+++ internal/initialize/templates/spectr/AGENTS.md.tmpl
