+++ internal/initialize/templates/spectr/instruction-pointer.md.tmpl
