+++ internal/initialize/templates/spectr/project.md.tmpl
