+++ internal/initialize/templates/tools/slash-apply.md.tmpl
