+++ internal/initialize/templates/tools/slash-proposal.md.tmpl
