+++ internal/initialize/templates/tools/slash-sync.md.tmpl
